Revert "Prereq requirements"
