+++ src/components/CourseView.tsx
@@ -3,7 +3,7 @@ import { Button, Col, Form, Row, Container } from "react-bootstrap";
 import { Course } from "../interfaces/course";
 import { DegreePlan } from "../interfaces/degreeplan";
 import { Semester } from "../interfaces/semester";
-import { CourseInformation } from "./CourseInformation";
+import { CoursePoolTable } from "./CoursePoolTable";
 /*this function generates the Container which displays all the course info, it 
     also contains the form for editing a course, and a lengthy function that will
     eventually call edit plan to process these edits and display the correct result*/
@@ -351,7 +351,7 @@ export function CourseView({
                     data-testid="editcourse-switch"
                 />
             </div>
-            <CourseInformation course={course}></CourseInformation>
+            <CoursePoolTable course={course}></CoursePoolTable>
         </Container>
     );
 }
