+++ src/components/PlanView.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { Button, Form } from "react-bootstrap";
 import { DegreePlan } from "../interfaces/degreeplan";
+import { CoursePool } from "./CoursePool";
 import { SemesterList } from "./SemesterList";
 import { CourseMover } from "./CourseMover";
 import { Course } from "../interfaces/course";
@@ -17,6 +18,7 @@ export function PlanView({
     editPlan: (id: string, newPlan: DegreePlan) => void;
     downloadPlan: (plan: DegreePlan) => void;
 }): JSX.Element {
+    const [showPool, setShowPool] = useState<boolean>(false);
     const [add, setAdd] = useState<boolean>(false);
     const [year, setYear] = useState<number>(0);
     const [session, setSession] = useState<string>("");
@@ -318,7 +320,10 @@ export function PlanView({
             setYear(inputToNumber);
         }
     }
-
+    //view CISC related courses in the pool
+    function showCoursePool() {
+        setShowPool(!showPool);
+    }
     //make semester list empty array
     function clearSemesters() {
         const newplan = {
@@ -562,7 +567,7 @@ export function PlanView({
                 </div>
             ) : null}
             <SemesterList plan={plan} editPlan={editPlan}></SemesterList>
-            {/* <div className="show-course-pool-button">
+            <div className="show-course-pool-button">
                 <Button
                     data-testid="show-pool-btn"
                     variant="success"
@@ -577,7 +582,7 @@ export function PlanView({
             </div>
             {showPool ? (
                 <CoursePool planpool={plan.planpool}></CoursePool>
-            ) : null} */}
+            ) : null}
         </div>
     );
 }
